refactor: simple menu types
